FerriteAssembly

Documentation for FerriteAssembly.

FerriteAssembly.assemble_cell!

FerriteAssembly.doassemble!

FerriteAssembly.doassemble!

FerriteAssembly.doassemble!

FerriteAssembly.scale_primary_global!

FerriteAssembly.scale_residual!

FerriteAssembly.unscale_primary!

FerriteAssembly.assemble_cell! — Method

assemble_cell!(
    assembler, cell, cellvalues, material, 
    state, anew, aold, dh_fh, Δt, cellcache)

Assemble the specific cell, provides a function barrier and uniform initialization of cellvalues, zero out element stiffness and residual, and scaling of primary and residual values

FerriteAssembly.doassemble! — Function

doassemble!(
    K::AbstractMatrix, r::AbstractVector, a::AbstractVector, 
    aold::AbstractVector, s::AbstractVector, dh::DofHandler, 
    cellvalues, material, Δt::Number, cache::CellCache, 
    cellset=nothing
    )

Assemble all cells by using the regular DofHandler dh. K, and r are the global stiffness matrix and residual vector to be calculated. a, and aoldare the current and old unknowns. s is a vector of state variables, where each element contains state variables for each cell. cellvalues are passed on to the element_routine after reinitialization. For multiple fields, it can be a Tuple or NamedTuple. material is a user defined type passed onto the element, and so is the time increment Δt. cache is described in CellCache.

FerriteAssembly.doassemble! — Function

doassemble!( K::AbstractMatrix, r::AbstractVector, anew::AbstractVector, aold::AbstractVector, state::Tuple, dh::MixedDofHandler, cellvalues::Tuple, materials::Tuple, Δt::Number, caches::Tuple, fullcellset=nothing )

Outer assembling loop using the MixedDofHandler dh

FerriteAssembly.doassemble! — Method

doassemble!(
    assembler, fh::FieldHandler, cellvalues, material, 
    s::AbstractVector, dh::MixedDofHandler, 
    anew::AbstractVector, aold::AbstractVector, 
    Δt::Number, cache::CellCache, fullcellset
    )

Inner assembling loop using the MixedDofHandler dh for a specific field fh. Normally not called by the user, only called from the other doassemble!(..., ::MixedDofHandler, ...)

FerriteAssembly.scale_primary_global! — Method

scale_primary_global!(a::AbstractVector, m::AbstractMaterial, dofs_tuple::NamedTuple)

Scale the global degrees of freedom a to normalize and make unitless

FerriteAssembly.scale_residual! — Method

scale_residual!(Ke::AbstractMatrix, re::AbstractVector, m::AbstractMaterial, dh_fh::Union{DofHandler, FieldHandler})

Scale the element stiffness, Ke, and residual, re to normalize wrt. units

FerriteAssembly.unscale_primary! — Method

unscale_primary!(ae::AbstractVector, m::AbstractMaterial, dh_fh::Union{DofHandler, FieldHandler})

Remove scaling factor on the element degrees of freedom ae