Implemented elements

There are two elements that are implemented directly in FerriteAssembly,

AbstractMaterial
If you have implemented your constitutive mechanical behavior following the MaterialModelsBase.jl's API, FerriteAssembly.element_routine! and FerriteAssembly.create_cell_state have already been implemented for this case.

WeakForm
For conveniently testing simple single field cases, an element where the weak form can be specified directly is implemented.

AbstractMaterial

FerriteAssembly.element_routine! — Method

FerriteAssembly.element_routine!(Ke, re, state, ae, m::MaterialModelsBase.AbstractMaterial, args...)

Solve the weak form

\[ \int_\Omega [\boldsymbol{\delta u}\otimes\nabla]^\mathrm{sym} : \boldsymbol{\sigma} \mathrm{d}\Omega = \int_\Gamma \boldsymbol{\delta u} \cdot \boldsymbol{t} \mathrm{d}\Gamma + \int_\Omega \boldsymbol{\delta u} \cdot \boldsymbol{b} \mathrm{d}\Omega\]

where $\sigma$ is calculated with the material_response function from MaterialModelsBase.jl. Note that create_cell_state is already implemented for <:AbstractMaterial.

WeakForm

FerriteAssembly.WeakForm — Type

WeakForm(f::Function)

Solve the problem with primary variable, $u$, variation, $\delta u$, and a weak form

\[ \int_\Omega f(\delta u, \delta u \otimes\nabla, u, u \otimes\nabla, \dot{u}, \dot{u} \otimes\nabla) \mathrm{d}\Omega - \int_\Gamma \delta u\ h(x,t,n)\ \mathrm{d}\Gamma - \int_\Omega \delta u\ b(x,t)\ \mathrm{d}\Omega = 0\]

where the function f is given to the weak form, and h and b are given with FerriteNeumann. This is only intended for testing, and is not optimized for speed (f is called for each shape function for each integration point)

Examples

Transient heat flow

Weak form

\[ \int_\Omega \delta u\ c\ \dot{u} - k\ [\nabla \delta u] \cdot [\nabla u]\ \mathrm{d}\Omega - \int_\Gamma \delta u\ q_\mathrm{n}\ \mathrm{d}\Gamma + \int_\Omega \delta u\ b\ \mathrm{d}\Omega = 0\]

Implementation

c = 1.0; k = 1.0; # heat capacity and heat conductivity (material parameters)
qn = 1.0; b=1.0;  # Normal boundary flux and internal heat source (external loading)
material = WeakForm((δu, ∇δu, u, ∇u, u_dot, ∇u_dot) -> δu*c*u_dot + k*(∇δu ⋅ ∇u))
nh = NeumannHandler(dh)
add!(nh, Neumann(:u, 2, getfaceset(dh.grid, "right"), (x,t,n)->qn))
add!(nh, BodyLoad(:c, 1, (x,t)->b))

Linear elasticity

Weak form

\[ \int_\Omega [\boldsymbol{\delta u}\otimes\nabla]^\mathrm{sym} : \boldsymbol{\sigma}\ \mathrm{d}\Omega - \int_\Gamma \boldsymbol{\delta u} \cdot \boldsymbol{t}\ \mathrm{d}\Gamma - \int_\Omega \boldsymbol{\delta u} \cdot \boldsymbol{b}\ \mathrm{d}\Omega = 0\]

Implementation

G = 80e3; K = 160e3; # Shear and bulk modulus (material parameters)
tn = 1.0, b=Vec((0.0, 0.0, -1.0)); # Normal traction and body force (external loading)
material = WeakForm((δu, ∇δu, u, ∇u, u_dot, ∇u_dot) -> (∇δu ⊡ (2*G*dev(symmetric(∇u)) + 3*K*vol(∇u)) )
nh = NeumannHandler(dh)
add!(nh, Neumann(:u, 2, getfaceset(dh.grid, "right"), (x,t,n)->tn*n))
add!(nh, BodyLoad(:c, 2, (x,t)->b))