Example elements

The package includes a set of example elements inside the submodule FerriteAssembly.ExampleElements. In order to use these it is possible to import them explicitly as, e.g.,

import FerriteAssembly.ExampleElements: StationaryFourier

Available example elements

StationaryFourier

TransientFourier

ElasticPlaneStress

PoroElasticPlaneStrain

FerriteAssembly.ExampleElements.StationaryFourier — Type

StationaryFourier(k)

For solving stationary linear heat conduction (which uses Fourier's law) with conductivity k, such that the heat flux is $\boldsymbol{q}=-k \nabla T$, where $T$ is the temperature field.

The strong form is,

\[ \nabla \cdot \boldsymbol{q} = h \quad \textbf{x} \in \Omega,\]

and the corresponding weak form is

\[ \int_\Omega [\nabla \delta T] \cdot \boldsymbol{q} \mathrm{d}\Omega = \int_\Gamma \delta T\ q_\mathrm{n} \mathrm{d}\Gamma - \int_\Omega \delta T\ h \mathrm{d}\Omega\]

where, on the right hand side, $q_\mathrm{n}$ is a heat flux normal to the boundary $\Gamma$, and $h$ is a volumetric heat supply. These contributions are not included in the element, and should be added with FerriteNeumann.jl

FerriteAssembly.ExampleElements.TransientFourier — Type

TransientFourier(k, c)

For solving the transient linear heat equation (which uses Fourier's law) with conductivity k, such that the heat flux is $\boldsymbol{q}=-k \nabla T$, where $T$ is the temperature field.

The strong form is,

\[ c\ \dot{T} + \nabla \cdot \boldsymbol{q} = h \quad \textbf{x} \in \Omega,\]

and the corresponding time-discretized weak form is

\[ \int_\Omega \delta T\ c\ \frac{T - {}^\mathrm{n}T}{\Delta t}\ \mathrm{d}\Omega + \int_\Omega [\nabla \delta T] \cdot \boldsymbol{q}\ \mathrm{d}\Omega = \int_\Gamma \delta T\ q_\mathrm{n}\ \mathrm{d}\Gamma - \int_\Omega \delta T\ h\ \mathrm{d}\Omega\]

where ${}^\mathrm{n}T$ is the old temperature (in the previous timestep) and $\Delta t$ is the timestep. On the right hand side, $q_\mathrm{n}$ is a heat flux normal to the boundary $\Gamma$, and $h$ is a volumetric heat supply. These external contributions on the right hand side are not included in the element, and should be added with FerriteNeumann.jl

FerriteAssembly.ExampleElements.ElasticPlaneStress — Type

ElasticPlaneStress(;E=2.e3, ν=0.3)

For solving linear elasticity for plane stress, where Young's modulus, E, and Poisson's ratio, ν, is used to construct the correct stiffness tensor, $\boldsymbol{\mathsf{C}}$, such that the stress, $\boldsymbol{\sigma}=\boldsymbol{\mathsf{C}}:\boldsymbol{\epsilon}$, where the strain tensor, $\boldsymbol{\epsilon}=[\boldsymbol{u}\otimes\nabla]^\mathrm{sym}$, is calculated from the displacement field, $\boldsymbol{u}(\boldsymbol{x},t)$.

The strong form of the mechanical quasi-static equilibrium is

\[ \boldsymbol{\sigma} \cdot \nabla + \boldsymbol{b} = 0 \quad \textbf{x} \in \Omega,\]

with the corresponding weak form,

\[ \int_\Omega [\boldsymbol{\delta u}\otimes\nabla]^\mathrm{sym} : \boldsymbol{\sigma} \mathrm{d}\Omega = \int_\Gamma \boldsymbol{\delta u} \cdot \boldsymbol{t} \mathrm{d}\Gamma + \int_\Omega \boldsymbol{\delta u} \cdot \boldsymbol{b} \mathrm{d}\Omega\]

The external loading on the right hand side is not included in the element, but can be implemented using FerriteNeumann.jl.

FerriteAssembly.ExampleElements.PoroElasticPlaneStrain — Type

PoroElasticPlaneStrain(;E=2.e3, ν=0.3, k=0.05, α=1.0, β=1/2e3)

TODO