MaterialModelsBase

If you have implemented your constitutive mechanical behavior following the MaterialModelsBase.jl's API, FerriteAssembly.element_routine! and FerriteAssembly.create_cell_state have already been implemented for this case.

FerriteAssembly.element_routine! — Method

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state_new,
    ae::AbstractVector, material, cellvalues, buffer)

The main function to be overloaded for the specific material. In most cases, the same implementation can be used for different cellvalues (e.g. for different interpolation orders) This function should modify the element stiffness matrix Ke, the residual re, and potentially state_new. The element degree of freedom values, ae, are filled by NaNs unless a is passed to doassemble!.

The following variables can be obtained from buffer.

get_state_old(buffer)

get_aeold(buffer)

get_time_increment(buffer)

dof_range(buffer, fieldname::Symbol)

getcoordinates(buffer)

celldofs(buffer)

cellid(buffer)

get_user_data(buffer)

get_cache(buffer)