Assemblers

FerriteAssembly.doassemble! — Function

doassemble!(assembler, new_states::Dict{Int}, buffer::AbstractDomainBuffer; a=nothing, aold=nothing, old_states=nothing, Δt=NaN)

Use assembler to assemble a single domain described by buffer, and update new_states if dictated by the assembler.

doassemble!(assembler, new_states::Dict{String}, buffers::Dict{String,AbstractDomainBuffer}; a=nothing, aold=nothing, old_states=nothing, Δt=NaN)

Use assembler to assemble all domains described by buffers, and update new_states if dictated by the assembler.

The keyword arguments work as follows:

a, aold: If these are nothing, then NaN values are given to the element routine for the corresponding entry.

old_states: If nothing, the old state in the cellbuffer will not be updated before calling the element routine

Δt: Directly the value returned by get_time_increment(cellbuffer)

Assembling system matrices and residual vectors

Element routines

One of the element methods should be overloaded for material. Note that cellvalues are already reinit!:ed for the current cell.

FerriteAssembly.element_routine! — Function

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state_new,
    ae::AbstractVector, material, cellvalues, buffer)

The main function to be overloaded for the specific material. In most cases, the same implementation can be used for different cellvalues (e.g. for different interpolation orders) This function should modify the element stiffness matrix Ke, the residual re, and potentially state_new. The element degree of freedom values, ae, are filled by NaNs unless a is passed to doassemble!.

The following variables can be obtained from buffer.

get_old_state(buffer)

get_aeold(buffer)

get_time_increment(buffer)

dof_range(buffer, fieldname::Symbol)

getcoordinates(buffer)

celldofs(buffer)

cellid(buffer)

get_user_data(buffer)

get_cache(buffer)

FerriteAssembly.element_routine!(Ke, re, state, ae, m::MaterialModelsBase.AbstractMaterial, args...)

Solve the weak form

\[ \int_\Omega [\boldsymbol{\delta u}\otimes\nabla]^\mathrm{sym} : \boldsymbol{\sigma}\ \mathrm{d}\Omega = \int_\Gamma \boldsymbol{\delta u} \cdot \boldsymbol{t}\ \mathrm{d}\Gamma + \int_\Omega \boldsymbol{\delta u} \cdot \boldsymbol{b}\ \mathrm{d}\Omega\]

where $\sigma$ is calculated with the material_response function from MaterialModelsBase.jl. Note that create_cell_state is already implemented for <:AbstractMaterial.

FerriteAssembly.element_residual! — Function

element_residual!(
    re::AbstractVector, state_new,
    ae::AbstractVector, material, cellvalues, buffer)

To calculate the element tangent stiffness Ke automatically by using ForwardDiff, it is possible to overload element_residual! instead of element_routine!. See element_routine! for a description of the input parameters.

!!! note MethodError with ForwardDiff.Dual When using automatic differentiation for elements with state variables (or other mutating values in e.g. cache), an error will be thrown if trying to change the type in many cases. When mutating new_state, call ForwardDiff.value() on the value to be assigned after it will no longer be used to calculate re. (If done on a value which is later affects re inside the element, the tangent, Ke, will be wrong.)

Assemblers

Ferrite.AssemblerSparsityPattern and Ferrite.AssemblerSymmetricSparsityPattern are supported. These are created as described in Ferrite.jl's documentation, and when passed to doassemble!, they assemble K and r as in Ferrite.jl. In addition, the following assemblers are provided by FerriteAssemble.jl,

FerriteAssembly.KeReAssembler — Type

KeReAssembler(K, r; fillzero=true, kwargs...)
KeReAssembler(a::Ferrite.AbstractSparseAssembler; kwargs...)

The default KeReAssembler works just like a Ferrite.AbstractSparseAssembler: It will call element_routine! and assemble Ke and re into the global system matrix K and vector r. However, it comes with the additional possible features, that are controllable via the keyword arguments:

ch::Union{Nothing,ConstraintHandler}=nothing: If a ConstraintHandler is given, local applications of constraints will be applied, using Ferrite.apply_assemble.

apply_zero: Required if a constraint handler is given, and forwarded to Ferrite.apply_assemble.

scaling: Calculate a scaling measure locally at the residual level, see e.g., ElementResidualScaling

a=Ferrite.start_assemble(K, r; fillzero=fillzero) is passed to the second definition if a matrix, K, and vector, r, are given as input.

FerriteAssembly.ReAssembler — Type

ReAssembler(r; fillzero=true, scaling=NoScaling())

ReAssembler will call element_residual! and assemble re into the system vector r. fillzero=true implies that r is zeroed upon construction of ReAssembler. Furthermore, remaining keyword arguments can enable the following features:

scaling: Calculate a scaling measure locally at the residual level, see e.g., ElementResidualScaling

Integration

In addition to assembling system matrices and vectors, much of the internal code can be reused to create quite efficient integration of values, given a solution vector (and potentially state variables) The general workflow assumes that setup_assembly has already been called, and that a solution vector is available. Then, it is possible to call doassemble! with an Integrator to obtain the integrated value. The following integrators are implemented

FerriteAssembly.Integrator — Type

Integrator(fun::Function, val)

Calculate the integral

\[\int_\Omega f(u, \nabla u, s)\, \mathrm{d}\Omega \]

For single-field problems, the function signature of $f$ is fun(u, ∇u, qp_state). u is the current function value, ∇u the current function gradient, and qp_state the current state in the current quadrature point. This assumes that cell_state::AbstractVector, otherwise, qp_state = cell_state.

For multi-field problem, we have fun(u::NamedTuple, ∇u::NamedTuple, qp_state), where the keys in u and ∇u are the fieldnames. The rest is same as for single-field problems.

It is the user's responsibility that fun(args...)::typeof(val). Additionally, the type of val must support

val + val

val * x (where x::Real)

zero(val)

One exception to these requirements; val::Tuple, if the elements of the tuple fulfills those requirements.

Example: Calculate the average value and gradient
We assume that we have done the setup: buffer, state = setup_assembly(...),
assembled K and r, and solved a=K\r

integrator = Integrator((u, ∇u, state)->(1.0, u, ∇u), (0.0, 0.0, zero(Vec{dim})))
doassemble!(integrator, states, buffer; a=a)
area = integrator.val[1]
u_avg = integrator.val[2]/area 
∇u_avg = integrator.val[3]/area

If the buffer is setup to be threaded, this calculation will also be threaded.