CellBuffer

A CellBuffer contains variables that are used and modified for each cell of a certain type. For each cell, it is reinit!:ed such that values such as cell coordinates, old state variables, etc. are updated to the current cell.

Construction

Construction of CellBuffer happens automatically when calling setup_assembly. To extra data can be included via the user_data keyword when calling setup_assembly or AssemblyDomain.

To allocate cache for the element routine, overload allocate_cell_cache:

FerriteAssembly.allocate_cell_cache — Method

FerriteAssembly.allocate_cell_cache(material, cellvalues)

This function can be overloaded for the specific material to allocate a cache that is stored in the CellBuffer which can be used to reduce allocations in the element routine. Returns nothing by default.

Access functions

The following access functions can be used to extract information from an AbstractCellBuffer.

FerriteAssembly.get_old_state — Function

get_old_state(c::CellBuffer)

Get the state variables for the cell from the previous time step.

Note

If no old_state keyword is passed to doassemble!, this variable will not be updated for the given cell, and typically contains the initial cell state.

FerriteAssembly.get_aeold — Function

FerriteAssembly.get_aeold(c::CellBuffer)

Get the old element dof-values for the current cell (Filled by NaNs unless aold is passed to doassemble!)

FerriteAssembly.get_time_increment — Function

get_time_increment(c::CellBuffer)

Get the time increment, Δt, that was passed to doassemble (defaults to NaN)

Ferrite.dof_range — Function

Ferrite.dof_range(c::CellBuffer, name::Symbol)

Get the dofrange::UnitRange{Int} for the dofs pertaining to the field: name. Same output as dof_range(dh::DofHandler, name), but fully type-stable.

Ferrite.getcoordinates — Function

Ferrite.getcoordinates(c::CellBuffer)

Ferrite.jl's getcoordinates function is overloaded on the CellBuffer to return the current cell's nodal coordinates.

Ferrite.celldofs — Function

Ferrite.celldofs(c::CellBuffer)

Ferrite.jl's celldofs function is overloaded on the CellBuffer to return the current cell's degree of freedom indices.

Ferrite.cellid — Function

Ferrite.cellid(c::CellBuffer)

Get the current cell id/nr

FerriteAssembly.get_user_data — Function

FerriteAssembly.get_user_data(c::CellBuffer)

Get the user specified user_data given to setup_assembly or AssemblyDomain

FerriteAssembly.get_cache — Function

FerriteAssembly.get_cache(c::CellBuffer)

Get the user-specified cache created by allocate_cell_cache