Residual scaling

There are many options for how to scale the residual in finite element simulations. This package does not intend to implement many different options, but does give the user the option to calculate scaling contributions from each cell, which may be useful. By defining a scaling that is passed to setup_assembly, it can be updated based on the output from each cell.

One type of scaling, ElementResidualScaling, is included as described below. Its code can be used as a template for how to include custom scaling that works on the element level.

ElementResidualScaling

FerriteAssembly.ElementResidualScaling — Type

ElementResidualScaling(dh::AbstractDofHandler, p=Val(2), T=Float64)

Create tolerance scaling based on the sum of the p-norm of each cell's residual vector, separately for each field. I.e. pseudo-code for field :u the scaling factor::T is

for each cell
    element_routine!(Ke, re, args...) # Calculate re
    factor += sum(abs.(re[dof_range(dh, :u)]).^p)^(1/p)

Note that p=Val(2) is a special case that is supported, for different values it should be given as a Real. p=2 is equivalent to Val(2), but is less efficient.