Setup API

FerriteAssembly.DomainSpec — Type

DomainSpec(sdh::SubDofHandler, material, fe_values; set=getcellset(sdh), colors_or_chunks=nothing, user_data=nothing)
DomainSpec(dh::DofHandler, material, fe_values; set=1:getncells(dh), colors=nothing, chunks=nothing, user_data=nothing)

Create a DomainSpec that can be used to set up a domain buffer. The element type of set determines the type of domain

Int: cell domain, fe_values should be CellValues (or a NamedTuple with CellValues as elements)

FaceIndex: face domain, fe_values should be FaceValues (or a NamedTuple with FaceValues as elements)

material is used for dispatch on the utilized worker's function. The user_data is just passed along (by reference) everywhere, and is accessible via the ItemBuffer (e.g. CellBuffer) given to the worker's function.

For multithreading, colors::Union{Vector{Vector{Int}},Vector{Vector{eltype(set)}} or chunks::Vector{Vector{Vector{eltype(set)}}} may be passed in order to partition the grid. Internally, if not given, chunks will be created from a colored grid (either given here, or calculated using Ferrite's default coloring algorithm). Note that the chunks must refer to the values in the set (exactly same), whereas colors can be for the entire grid, although it is usually better to color each cellset individually.

FerriteAssembly.setup_domainbuffer — Function

setup_domainbuffer(domain::DomainSpec; a=nothing, threading=false, autodiffbuffer=false)

Setup a domain buffer for a single grid domain, domain.

a::Vector: The global degree of freedom values are used to pass the local element dof values to the create_cell_state function, making it possible to create the initial state dependent on the initial conditions for the field variables.

threading: Should a ThreadedDomainBuffer be created to work the grid multithreaded if supported by the used worker?

autodiffbuffer: Should a custom itembuffer be used to speed up the automatic differentiation (if supported by the itembuffer)

FerriteAssembly.setup_domainbuffers — Function

setup_domainbuffers(domains::Dict{String,DomainSpec}; kwargs...)

Setup multiple domain buffers, one for each DomainSpec in domains. See setup_domainbuffer for description of the keyword arguments.

AbstractDomainBuffer

The domain buffer be a DomainBuffer, ThreadedDomainBuffer, or a Dict{String} with eltype of one of the former. The following functions are defined for these buffers:

FerriteAssembly.get_material — Method

get_material(dbs::Dict{String,AbstractDomainBuffer}, domain::String)
get_material(db::AbstractDomainBuffer)

Get the material for the domain represented by db or dbs[domain].

FerriteAssembly.get_dofhandler — Method

get_dofhandler(dbs::Dict{String,AbstractDomainBuffer})
get_dofhandler(db::AbstractDomainBuffer)

Get the dofhandler stored in db. Note that this is the global dofhandler, and not the SubDofHandler that is local to a specific domain.

FerriteAssembly.get_state — Method

get_state(dbs::Dict{String,AbstractDomainBuffer}, domain::String)
get_state(db::AbstractDomainBuffer)

Get the states::Dict{Int,S}, where S type of the state for each entity in the domain, stored in the db or dbs[domain].

FerriteAssembly.get_old_state — Method

get_old_state(dbs::Dict{String,AbstractDomainBuffer}, domain::String)
get_old_state(db::AbstractDomainBuffer)

Get the states::Dict{Int,S}, where S type of the state for each entity in the domain, stored in the db or dbs[domain].

FerriteAssembly.update_states! — Method

update_states!(db::Dict{String,AbstractDomainBuffer})
update_states!(db::AbstractDomainBuffer)

Update the states such that old_states = states for the states stored in db.

This method tries to avoid allocating new values where possible. Currently, if create_cell_state returns T or Vector{T} where isbitstype(T), this works. If needed/wanted, it should be relatively easy to provide an interface to make it possible to have allocation free for custom cell states.

FerriteAssembly.set_time_increment! — Method

set_time_increment!(db::Dict{String,AbstractDomainBuffer}, Δt)
set_time_increment!(db::AbstractDomainBuffer, Δt)

Update the time increment stored in db, which is passed on to the stored AbstractItemBuffer

Ferrite.getcellset — Method

getcellset(dbs::Dict{String,AbstractDomainBuffer}, domain::String)
getcellset(db::AbstractDomainBuffer)

Get the cellset for db or dbs[domain]