State variables

The initial state variables may vary depending on the position in the grid. Furthermore, the datastructure depends on the type of dof handler, so a convenience function exists that creates the correct variable structure.

FerriteAssembly.create_cell_state — Function

create_cell_state(material, cellvalues, x, ae, dofrange)

Defaults to returning nothing.

Overload this function to create the state which should be passed into the element_routine!/element_residual! for the given material and cellvalues. x is the cell's coordinates, ae the element degree of freedom values, and dofrange::NamedTuple containing the local dof range for each field. As for the element routines, ae, is filled with NaN unless the global degree of freedom vector is given to the setup_assembly function.

FerriteAssembly.update_states! — Function

update_states!(old_states, new_states)

Update old_states with the values from new_states. This is typically done after a converged time increment.

This method tries to avoid allocating new values where possible. Currently, if create_cell_state returns T or Vector{T} where isbitstype(T), this works.

If needed/wanted, it should be relatively easy to provide an interface to make it possible to have allocation free for custom cell states.

FerriteAssembly.create_states — Function

create_states(sdh::SubDofHandler, material, cellvalues, a, cellset, dofrange)

Create a Dict of states for the cells in cellset, where the user should define the create_cell_state function for their material (and corresponding cellvalues) dofrange::NamedTuple is passed onto create_cell_state and contains the local dof ranges for each field.