Workers

The to perform work, such as assembling, the function work! should be called with a given worker and domain buffer.

There are currently two categories of workers implemented: which are used to calculate

Assemblers: Calculate system matrices and vectors

Integrators: Integrate a function over the domain

FerriteAssembly.work! — Function

work!(worker, buffer; a=nothing, aold=nothing)

Perform the work according to worker over the domain(s) specificed by buffer. Current, a, and old, aold, global degree of freedom vectors are passed to get these values passed into the innermost user-defined functions. If not passed (or as nothing), NaN values are passed into the innermost functions.

System matrix and vector

In order to assemble the system matrices and residual vectors, it is necessary to define the appropriate element routine and choose an assembler. The following assemblers can be used to assemble the system matrix and vector:

Ferrite.start_assemble that returns

Ferrite.AssemblerSparsityPattern

Ferrite.AssemblerSymmetricSparsityPattern

KeReAssembler

ReAssembler

Where the ones available in FerriteAssembly have additional features, such as the possibility of applying constraints locally (see Ferrite.apply_assemble!) or calculate scaling for the residual.

Element routines

One of the element methods should be overloaded for material. Note that cellvalues are already reinit!:ed for the current cell.

FerriteAssembly.element_routine! — Method

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state,
    ae::AbstractVector, material, cellvalues, buffer)

The main function to be overloaded for the specific material. In most cases, the same implementation can be used for different cellvalues (e.g. for different interpolation orders) This function should modify the element stiffness matrix Ke, the residual re, and potentially state. The element degree of freedom values, ae, are filled by NaNs unless a is passed to work!.

The following variables can be obtained from buffer.

get_old_state(buffer)

get_aeold(buffer)

get_time_increment(buffer)

dof_range(buffer, fieldname::Symbol)

getcoordinates(buffer)

celldofs(buffer)

cellid(buffer)

get_user_data(buffer)

get_user_cache(buffer)

FerriteAssembly.element_residual! — Function

element_residual!(
    re::AbstractVector, state,
    ae::AbstractVector, material, cellvalues, buffer)

To calculate the element tangent stiffness Ke automatically by using ForwardDiff, it is possible to overload element_residual! instead of element_routine!. See element_routine! for a description of the input parameters.

Note

MethodError with ForwardDiff.Dual
When using automatic differentiation for elements with state variables (or other mutating values in e.g. cache), an error will be thrown if trying to change the type in many cases. When mutating state, call ForwardDiff.value() on the value to be assigned after it will no longer be used to calculate re. (If done on a value which is later affects re inside the element, the tangent, Ke, will be wrong.)

Assemblers

FerriteAssembly.KeReAssembler — Type

KeReAssembler(K, r; fillzero=true, kwargs...)
KeReAssembler(a::Ferrite.AbstractSparseAssembler; kwargs...)

The default KeReAssembler works just like a Ferrite.AbstractSparseAssembler: It will call element_routine! and assemble Ke and re into the global system matrix K and vector r. However, it comes with the additional possible features, that are controllable via the keyword arguments:

ch::Union{Nothing,ConstraintHandler}=nothing: If a ConstraintHandler is given, local applications of constraints will be applied, using Ferrite.apply_assemble.

apply_zero: Required if a constraint handler is given, and forwarded to Ferrite.apply_assemble.

scaling: Calculate a scaling measure locally at the residual level, see e.g., ElementResidualScaling

a=Ferrite.start_assemble(K, r; fillzero=fillzero) is passed to the second definition if a matrix, K, and vector, r, are given as input.

reset_scaling! is called when creating KeReAssembler

FerriteAssembly.ReAssembler — Type

ReAssembler(r; fillzero=true, scaling=NoScaling())

ReAssembler will call element_residual! and assemble re into the system vector r. fillzero=true implies that r is zeroed upon construction of ReAssembler. Furthermore, remaining keyword arguments can enable the following features:

scaling: Calculate a scaling measure locally at the residual level, see e.g., ElementResidualScaling.

reset_scaling! is called when creating ReAssembler

Integration

In addition to assembling system matrices and vectors, much of the internal code can be reused to create quite efficient integration of values, given a solution vector (and potentially state variables). The same domain buffer as for assembly can be used, and exactly as for assembly, we just call work!(integrator, buffer). The following integrators are implemented

FerriteAssembly.SimpleIntegrator — Type

SimpleIntegrator(fun::Function, val)

Calculate the integral

\[\int_\Omega f(u, \nabla u, s)\, \mathrm{d}\Omega \]

For single-field problems, the function signature of $f$ is fun(u, ∇u, qp_state). u is the current function value, ∇u the current function gradient, and qp_state the current state in the current quadrature point. This assumes that cell_state::AbstractVector, otherwise, qp_state = cell_state.

For multi-field problem, we have fun(u::NamedTuple, ∇u::NamedTuple, qp_state), where the keys in u and ∇u are the fieldnames. The rest is same as for single-field problems.

It is the user's responsibility that fun(args...)::typeof(val). Additionally, the type of val must support

val + val

val * x (where x::Real)

zero(val)

One exception to these requirements; val::Tuple, if the elements of the tuple fulfills those requirements.

Specify domains to only integrate over a part of the grid. It should contain a subset of the keys provided to setup_domainbuffers. Single, String, or collections; Set{String}, AbstractVector{String}, or NTuple{N,String} inputs are supported. If domains=nothing, all domains are integrated.

Example: Calculate the average value and gradient
We assume that we have done the setup: buffer = setup_domainbuffer(...),
assembled K and r, and solved a=K\r

integrator = SimpleIntegrator((u, ∇u, state)->(1.0, u, ∇u), (0.0, 0.0, zero(Vec{dim})))
work!(integrator, buffer; a=a)
area = integrator.val[1]
u_avg = integrator.val[2]/area 
∇u_avg = integrator.val[3]/area

If the buffer is setup to be threaded, this calculation will also be threaded.

FerriteAssembly.Integrator — Type

Integrator(val::V; domains=nothing)

Integrate over the domain by modifying the value val for each cell in the function integrate_cell, which should be overloaded for the specific combination of val::V and the material for that cell. V must be a mutable type.

Specify domains to only integrate over a part of the grid. It should contain the names of the domains to integrate over. Single, String, or collections; Set{String}, AbstractVector{String}, or NTuple{N,String} inputs are supported. If domains=nothing, all domains are integrated.

Note: This integrator will currently only run sequentially even if a threaded assembly is setup with setup_domainbuffer

Integration routines

The general integrator requires overloading the integrate_cell! function

FerriteAssembly.integrate_cell! — Function

integrate_cell!(val, cell_state, ae, material, cv, cellbuffer)

Function to be overloaded for val (potentially in combination with material), and will be called when using Integrator. Mutate val to add to the result.

The order of the inputs is chosen to follow the element routines