API

Methods that should be overloaded and exported functions are described on this page.

Setting up and doing the assembly

FerriteAssembly.setup_assembly — Function

setup_assembly(dh, material, cellvalues; kwargs...)

Setup assembly for a single domain (i.e. the same material and interpolations everywhere). Returns the buffer, old_states, and new_states to be used in doassemble!. The state variables are created by calling the user-defined create_cell_state function. Available keyword arguments

a=nothing: Give the global dof vector to pass element dofs, ae, to create_cell_state (NaN-values otherwise)

colors=nothing: Give colors for the grid from Ferrite.create_coloring to setup threaded assembly. If nothing, the assembly is sequential.

autodiffbuffer=Val(false): Set to true or Val(true) (for type stable construction) to use AutoDiffCellBuffer instead of CellBuffer.

user_data=nothing: The user_data is passed to each AbstractCellBuffer by reference (when threaded)

cache=nothing: The cache is passed to each AbstractCellBuffer, and deepcopied if threaded.

scaling=nothing: The scaling to be calculated, see e.g. ElementResidualScaling

setup_assembly(domains::Vector{<:AssemblyDomain}; kwargs...)

Setup assembly for each AssemblyDomain in domains. Returns the Dicts buffers, old_states, and new_states to be used in doassemble!. The state variables are created by calling the user-defined create_cell_state function. Available keyword arguments

a=nothing: Give the global dof vector to pass element dofs, ae, to create_cell_state (NaN-values otherwise)

colors=nothing: Give colors for the grid from Ferrite.create_coloring to setup threaded assembly. If nothing, the assembly is sequential.

autodiffbuffer=Val(false): Set to true or Val(true) (for type stable construction) to use AutoDiffCellBuffer instead of CellBuffer.

scaling=nothing: The scaling to be calculated, see e.g. ElementResidualScaling

FerriteAssembly.AssemblyDomain — Type

AssemblyDomain(name, dh, material, cellvalues; cellset, user_data=nothing, cache=nothing)

Create an AssemblyDomain that can be used when calling setup_assembly to assemble multiple domains. name is used to access the corresponding DomainBuffer and state variables returned by setup_assembly. If not given, cellset is attempted to be inferred from the DofHandler, dh.

FerriteAssembly.doassemble! — Function

doassemble!([K::AbstractMatrix,] r::AbstractVector, new_states, old_states, buffer; kwargs...)

Assemble K and r (element_residual must be defined to only assemble r), given the new_states, old_states, and buffer returned from setup_assembly. The available keyword arguments are

a=nothing: The current degree of freedom vector. If nothing, ae with NaN values is passed to element_routine!

aold=nothing: The old degree of freedom vector. If nothing, get_aeold(buffer) gives aNaN`-filled vector.

Δt=NaN: The time increment that can be accessed as get_time_increment(buffer) in the element routines.

fillzero=true: Should K and r be zeroed before starting the assembly?

Element routines

One of the element methods should be overloaded for a given combination of cellvalues and material. Note that the cellvalues are already reinit!:ed when passed to the element routines.

FerriteAssembly.element_routine! — Function

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state_new,
    ae::AbstractVector, material, cellvalues, buffer)

The main function to be overloaded for the specific material. In most cases, the same implementation can be used for different cellvalues (e.g. for different interpolation orders) This function should modify the element stiffness matrix Ke, the residual re, and potentially state_new. The element degree of freedom values, ae, are filled by NaNs unless a is passed to doassemble!.

The following variables can be obtained from buffer.

get_state_old(buffer)

get_aeold(buffer)

get_time_increment(buffer)

dof_range(buffer, fieldname::Symbol)

getcoordinates(buffer)

celldofs(buffer)

cellid(buffer)

get_user_data(buffer)

get_cache(buffer)

FerriteAssembly.element_routine!(Ke, re, state, ae, m::MaterialModelsBase.AbstractMaterial, args...)

Solve the weak form

\[ \int_\Omega [\boldsymbol{\delta u}\otimes\nabla]^\mathrm{sym} : \boldsymbol{\sigma}\ \mathrm{d}\Omega = \int_\Gamma \boldsymbol{\delta u} \cdot \boldsymbol{t}\ \mathrm{d}\Gamma + \int_\Omega \boldsymbol{\delta u} \cdot \boldsymbol{b}\ \mathrm{d}\Omega\]

where $\sigma$ is calculated with the material_response function from MaterialModelsBase.jl. Note that create_cell_state is already implemented for <:AbstractMaterial.

FerriteAssembly.element_residual! — Function

element_residual!(
    re::AbstractVector, state_new,
    ae::AbstractVector, material, cellvalues, buffer)

To calculate the element tangent stiffness Ke automatically by using ForwardDiff, it is possible to overload element_residual! instead of element_routine!. See element_routine! for a description of the input parameters.

!!! note MethodError with ForwardDiff.Dual When using automatic differentiation for elements with state variables (or other mutating values in e.g. cache), an error will be thrown if trying to change the type in many cases. When mutating new_state, call ForwardDiff.value() on the value to be assigned after it will no longer be used to calculate re. (If done on a value which is later affects re inside the element, the tangent, Ke, will be wrong.)

State variables

The initial state variables may vary depending on the position in the grid. Furthermore, the datastructure depends on the type of dof handler, so a convenience function exists that creates the correct variable structure.

FerriteAssembly.create_cell_state — Function

create_cell_state(material, cellvalues, x, ae, dofrange)

Defaults to returning nothing.

Overload this function to create the state which should be passed into the element_routine!/element_residual! for the given material and cellvalues. x is the cell's coordinates, ae the element degree of freedom values, and dofrange::NamedTuple containing the local dof range for each field. As for the element routines, ae, is filled with NaN unless the global degree of freedom vector is given to the setup_assembly function.

FerriteAssembly.update_states! — Function

update_states!(old_states, new_states)

Update old_states with the values from new_states. This is typically done after a converged time increment.

This method tries to avoid allocating new values where possible. Currently, if create_cell_state returns T or Vector{T} where isbitstype(T), this works.

If needed/wanted, it should be relatively easy to provide an interface to make it possible to have allocation free for custom cell states.

FerriteAssembly.create_states — Function

create_states(sdh::SubDofHandler, material, cellvalues, a, cellset, dofrange)

Create a Dict of states for the cells in cellset, where the user should define the create_cell_state function for their material (and corresponding cellvalues) dofrange::NamedTuple is passed onto create_cell_state and contains the local dof ranges for each field.

CellBuffer

Variables that are used and modified for each cell of a certain type, but that don't belong to a specific cell, are collected in a CellBuffer.

Access functions

The following access functions can be used to extract information from the CellBuffer

FerriteAssembly.get_state_old — Function

get_state_old(c::CellBuffer)

Get the state variables for the cell from the previous time step.

FerriteAssembly.get_aeold — Function

FerriteAssembly.get_aeold(c::CellBuffer)

Get the old element dof-values for the current cell (Filled by NaNs unless aold is passed to doassemble!)

FerriteAssembly.get_time_increment — Function

get_time_increment(c::CellBuffer)

Get the time increment, Δt, that was passed to doassemble

Ferrite.dof_range — Function

Ferrite.dof_range(c::CellBuffer, name::Symbol)

Get the dofrange::UnitRange{Int} for the dofs pertaining to the field: name. Same output as dof_range(dh::DofHandler, name), but fully type-stable.

Ferrite.getcoordinates — Function

Ferrite.getcoordinates(c::CellBuffer)

Ferrite.jl's getcoordinates function is overloaded on the CellBuffer to return the current cell's nodal coordinates.

Ferrite.celldofs — Function

Ferrite.celldofs(c::CellBuffer)

Ferrite.jl's celldofs function is overloaded on the CellBuffer to return the current cell's degree of freedom indices.

Ferrite.cellid — Function

Ferrite.cellid(c::CellBuffer)

Get the current cell id/nr

FerriteAssembly.get_user_data — Function

FerriteAssembly.get_user_data(c::CellBuffer)

Get the user specified user_data given to CellBuffer

FerriteAssembly.get_cache — Function

FerriteAssembly.get_cache(c::CellBuffer)

Get the user-specified cache given to the CellBuffer

Residual scaling

There are many options for how to scale the residual in finite element simulations. This package does not intend to implement many different options, but does give the user the option to calculate scaling contributions from each cell, which may be useful. By defining a scaling that is passed to setup_assembly, it can be updated based on the output from each cell.

One type of scaling, ElementResidualScaling, is included as described below. Its code can be used as a template for how to include custom scaling that works on the element level.

In general, a scaling must support the following functions

FerriteAssembly.update_scaling! — Function

update_scaling!(scaling, re, cellbuffer)

This function should add the contribution from the element residual vector re to the scaling factors in scaling.

FerriteAssembly.reset_scaling! — Function

reset_scaling!(scaling)

This function should reset the scaling factors, such that the values don't accumulate if used in multiple iterations/time steps.

FerriteAssembly.add_to_scaling! — Function

add_to_scaling!(base_scaling, scaling)

Add the scaling values from scaling to base_scaling. This function is used after threaded assembly to put all the scaling values together into the object initially sent to setup_assembly

ElementResidualScaling

FerriteAssembly.ElementResidualScaling — Type

ElementResidualScaling(dh::AbstractDofHandler, p=Val(2), T=Float64)

Create tolerance scaling based on the sum of the p-norm of each cell's residual vector, separately for each field. I.e. pseudo-code for field :u the scaling factor::T is

for each cell
    element_routine!(Ke, re, args...) # Calculate re
    factor += sum(abs.(re[dof_range(dh, :u)]).^p)^(1/p)

Note that p=Val(2) is a special case that is supported, for different values it should be given as a Real. p=2 is equivalent to Val(2), but is less efficient.