Hacking API

The functions described on this page are somewhere inbetween regular API and internal API, and are intended for making custom solutions, but with a higher risk of breaking changes. Breaking changes to these interfaces should still be followed by a version bump.

TaskLocals API (task local storage)

During multithreaded assembly, each task needs its own storage with values that it can change. This can be both cache variables (whose values don't matter after the task completes) or other values that are part of the result from the assembly procedure. To do this in a structured way, the TaskLocals type and associated interface is used for all these cases in the package, but the user should never "see" this type directly (but it can be embedded in types seen by the user, such as ThreadedDomainBuffer). Specifically, a so-called scatter-gather approach is emulated, even though the memory is shared, allowing this to be simplified.

FerriteAssembly.TaskLocals — Type

TaskLocals(base; num_tasks=Threads.nthreads())

TaskLocals is used to store a base value and a vector, locals. Items in locals are for each task and can be obtained with get_local. base can be obtained with get_base.

A base-type that supports the TaskLocals API should define the following functions

create_local

scatter!(local, base)

gather!(base, local)

FerriteAssembly.create_local — Function

create_local(base::TB)

Create a task local variable, local::TL. In many cases, TL=TB. local's state should match that after calling gather!(base, local) (reset local) followed by scatter!(local, base) (add info from base) (This is important, because it allows create_local to be called while working with tasks, if more locals are needed)

FerriteAssembly.scatter! — Function

scatter!(local, base)

Write any information from base required to be forwarded to the local::typeof(create_local(base)). (typically if something in base have changed since its creation)

FerriteAssembly.gather! — Function

gather!(base, local)

Take any information from local::typeof(create_local(base)) that should be added to base after an assembly. In addition, any accumulative information in local should be reset, such that running scatter! -> "do work" -> gather! should only affect the values in base.

FerriteAssembly.get_local — Function

get_local(tl::TaskLocals, i::Int)

Get the ith local variable from tl

FerriteAssembly.get_base — Function

get_base(tl::TaskLocals)

Get the base variable from tl

Assembler interface

Different types of assemblers can be created in addition to those already defined by the package. The interface for creating an assembler is that the assembler must support the TaskLocals API, as well as the method, FerriteAssembly.assemble_cell_reinited!(assembler, cell_state, buffer::AbstractCellBuffer), where cell_state is the output from create_cell_state. The following methods for builtin assemblers can be used as examples

FerriteAssembly.assemble_cell_reinited! — Function

assemble_cell_reinited!(assembler::Ferrite.AbstractSparseAssembler, cell_state, cellbuffer)

Calculate the local element stiffness, Ke, and the local residual, re, with element_routine! for the cell described by the reinitialized cellbuffer. Assemble these into the global matrix and vector in assembler.

assemble_cell_reinited!(assembler::ReAssembler, cell_state, cellbuffer)

Calculate the local residual, re, with element_residual! for the cell described by the reinitialized cellbuffer. Assemble re into the global vector in assembler. In addition, ReAssembler supports the extra option

Residual scaling factor, e.g. ElementResidualScaling

assemble_cell_reinited!(assembler::KeReAssembler, cell_state, cellbuffer)

Calculate the local element stiffness, Ke, and the local residual, re, with element_routine! for the cell described by the reinitialized cellbuffer. Assemble these into the global matrix and vector in assembler. In addition, KeReAssembler supports the extra options

Local application of constraints: Ferrite.apply_assemble!

Residual scaling factor, e.g. ElementResidualScaling

In addition, the following functions are used to determine properties/requirements for the assembler.

can_thread(assembler)::Bool: Is multithreading supported, if false, sequential assembly will run even if the user has set up multithreaded assembly with setup_assembly. If not defined, defaults to false.

needs_colors(assembler)::Bool: Indicate whether the assembler needs to run in colored mode or not, defaults to true. (Currently not used, all multithreading runs colored, but may be changed.)

skip_this_domain(assembler, name::String)::Bool: Should the domain with name (given to AssemblyDomain) be skipped (currently used by the integrators to allow integrating only part of the domain). Defaults to false.

Custom scaling

In general, a scaling must, in addition to the TaskLocals API, support the following functions

FerriteAssembly.update_scaling! — Function

update_scaling!(scaling, re, cellbuffer)

This function should add the contribution from the element residual vector re to the scaling factors in scaling.

FerriteAssembly.reset_scaling! — Function

reset_scaling!(scaling)

This function should reset the scaling factors, such that the values don't accumulate if used in multiple iterations/time steps.