Scaling

Note that this part just exposes temporary docstrings and the interface might change.

Reminder: Might make sense to rename "primary" to "unknowns"? This would go better along with FESolvers

FerriteAssembly.scale_primary_global! — Function

scale_primary_global!(a::AbstractVector, material, dofs_tuple::NamedTuple)

Scale the global degrees of freedom a to normalize and make unitless

FerriteAssembly.unscale_primary_global! — Function

unscale_primary_global!(a, material, dofs_tuple::NamedTuple)

Unscale the global degrees of freedom a to get correct values for e.g. postprocessing.

FerriteAssembly.unscale_primary! — Function

unscale_primary!(ae::AbstractVector, m::AbstractMaterial, dh_fh::Union{DofHandler, FieldHandler})

Remove scaling factor on the element degrees of freedom ae

FerriteAssembly.scale_residual! — Function

scale_residual!(Ke::AbstractMatrix, re::AbstractVector, m::AbstractMaterial, dh_fh::Union{DofHandler, FieldHandler})

Scale the element stiffness, Ke, and residual, re to normalize wrt. units