FerriteAssembly

The goal of FerriteAssembly is to provide a simple structure to perform assembly in Ferrite.jl.

Both the DofHandler and MixedDofHandler are supported. Currently only sequential assembly, but this will be extended to threaded assembly.

The package work by exporting the doassemble! function, and requires the user to define either element_routine! (calculate both Ke and re), or just element_residual! (calculate only re). In the latter case, Ke is calculated by ForwardDiff.jl

An advanced option to scale the unknowns, residual, and jacobian exists (currently undocumented)

FerriteAssembly

A first example

Detailed API description

Internal API

Scaling

A first example

We take a slightly modified element routine from Ferrite.jl's heat equation example. We assume that we already have defined dh::DofHandler and our cellvalues::CellScalarValues according to that example. We then define

struct ThermalMaterial end
material = ThermalMaterial()

and our element_routine! for that material as

function FerriteAssembly.element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, 
    ae_new::AbstractVector, ae_old::AbstractVector,
    state::AbstractVector, material::ThermalMaterial, 
    cellvalues::CellScalarValues, 
    dh_fh::Union{DofHandler,FieldHandler}, Δt, materialcache
    )
    n_basefuncs = getnbasefunctions(cellvalues)
    # Loop over quadrature points
    for q_point in 1:getnquadpoints(cellvalues)
        dΩ = getdetJdV(cellvalues, q_point)
        for i in 1:n_basefuncs
            δu  = shape_value(cellvalues, q_point, i)
            ∇δu = shape_gradient(cellvalues, q_point, i)
            # Add body load contribution to re
            re[i] += -δu * dΩ
            # Loop over trial shape functions
            for j in 1:n_basefuncs
                ∇u = shape_gradient(cellvalues, q_point, j)
                # Add contribution to Ke
                Ke[i, j] += (∇δu ⋅ ∇u) * dΩ
            end
        end
    end
end

We can now create the global vectors and matrices, as well as the time step

K = create_sparsity_pattern(dh)
a=zeros(ndofs(dh)); aold=copy(a);
r = zeros(ndofs(dh))
Δt=1.0

Before we can assemble, we need our cellcache (to avoid unecessary allocations inside the assembly loop). But this is created automatically from the DofHandler:

cache = CellCache(dh)

And then we can assemble our system by calling

doassemble!(K, r, a, aold, states, dh, cv, material, Δt, cache)

Detailed API description

One of the element methods should be overloaded for a given combination of cellvalues and material.

FerriteAssembly.element_routine! — Function

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, 
    ae::AbstractVector, ae_old::AbstractVector,
    state, material, cellvalues, 
    dh_fh::Union{DofHandler,FieldHandler}, Δt, materialcache
    )

The main function to be overloaded for the specific material and cellvalues. This function should modify the residual r and the element stiffness matrix Ke, such that Ke=∂(re)/∂(ae), where ae are the element degrees of freedom. ae_old are the corresponding old values

state should contain a state description for the element. Typically, state will be a vector with a state variable for each integration point, but it can also be any other type for each element. On input, these are the old values and should be mutated to the updated value for the current time step.

The user defined material variable usually contain the material parameters

cellvalues should contain the cell values for the given element. It can also be a tuple or named tuple of cellvalues.

When the regular DofHandler is used, dh_fh::DofHandler is passed to the element routine. However, if the MixedDofHandler is used, one of its fieldhandlers are passed as dh_fh::FieldHandler. This gives the option to call dof_range(dh_fh, field::Symbol) for multi-field problems.

Finally, a time increment Δt and an optional materialcache are passed into the element as well.

FerriteAssembly.element_residual! — Function

element_residual!(
    re::AbstractVector, 
    ae::AbstractVector, ae_old::AbstractVector,
    state, material, cellvalues, 
    dh_fh::Union{DofHandler,FieldHandler}, Δt, materialcache
    )

To calculate the element tangent stiffness Ke automatically by using ForwardDiff, it is possible to overload element_residual! instead of element_routine!. See element_routine! for a description of the input parameters.

Note that in order for this function to work, care must be taken when mutating values to not change their types. When mutating the state, ensure to call ForwardDiff.value() on those values. Warning: Only do this at a point when the calculation of re is unaffected by the values in state, otherwise Ke will be wrong.

The allocations are reduced by saving all variables to a cellcache, which is automatically created by giving it the dof handler:

FerriteAssembly.CellCache — Type

CellCache(n::Int, materialcache=nothing)

Create a cell cache for an element with n degrees of freedom, as well as with materialcache

CellCache(dh::DofHandler, materialcache=nothing)

Create a cell cache for an element with n=ndofs_per_cell(dh), as well as with materialcache.

CellCache(dh::MixedDofHandler, fh::FieldHandler, materialcache=nothing)

Create a cell cache for an element with n degrees of freedom, as well as with materialcache. Here, n is the number of degrees of freedom for the cells in fh.cellset.

CellCache(dh::MixedDofHandler, materialcache=nothing)
CellCache(dh::MixedDofHandler, materialcache::Tuple)

Returns a tuple of CellCaches for each FieldHandler in dh.fieldhandlers. If materialcache::Tuple is given, materialcache[i] is assigned to the CellCache corresponding to dh.fieldhandlers[i]. Otherwise, materialcache is assigned to each CellCache

Once everything is set up, one can call the function which will actually do the assembly:

FerriteAssembly.doassemble! — Function

doassemble!(
    K::AbstractMatrix, r::AbstractVector, a::AbstractVector, 
    aold::AbstractVector, s::AbstractVector, dh::DofHandler, 
    cellvalues, material, Δt::Number, cache::CellCache
    )

Assemble all cells by using dh::DofHandler.

K, and r are the global stiffness matrix and residual vector to be calculated.

a and aold are the current and old unknowns.

s is a vector of state variables, where each element contains state variables for each cellid. If cellset is given (see below), the number is still the cellid, and in some cases it might make sense to pass a sparse vector.

cellvalues are passed on to the element routine after calling Ferrite.reinit!. (Multiple cellvalues can be given as a Tuple or NamedTuple, and all elements will be reinitialized)

The user-defined material and the time increment Δt are passed into the element routine.

cache is described in CellCache.

doassemble!(
    K::AbstractMatrix, r::AbstractVector, anew::AbstractVector, 
    aold::AbstractVector, state::Tuple, dh::MixedDofHandler, 
    cellvalues::Tuple, materials::Tuple, Δt::Number, caches::Tuple
    )

Assemble all cells by using dh::MixedDofHandler. In this case, we first loop over each fieldhandler in dh. Therefore, some variables must now be given as tuples, where each element corresponds to the values in the specific fieldhandler.

K, and r are the global stiffness matrix and residual vector to be calculated.

a and aold are the current and old unknowns.

s is a tuple of vectors of state variables. Each element, s[i], corresponds to the ith element in dh.fieldhandlers. The jth element of s[i]::AbstractVector, contains the state variables for cellid=j. If the state-type differs for multiple fieldhandlers, it might make sense to use sparse vectors.

cellvalues is a tuple with one element for each fieldhandler. For each fieldhandler i, cellvalues[i] is passed to the element routine after calling Ferrite.reinit!. (Multiple cellvalues can be given by making cellvalues[i] a Tuple or NamedTuple. All elements will be reinitialized)

The user-defined material and the time increment Δt are passed into the element routine. Finally,

caches is a tuple with CellCache-elements, which are described in CellCache.