FerriteAssembly

The goal of FerriteAssembly is to provide a simple structure to perform assembly in Ferrite.jl.

Sequential and threaded assembly of both the DofHandler and MixedDofHandler are supported.

The package works by exporting the doassemble! function, and require the user to define either element_routine! (calculate both Ke and re), or just element_residual! (calculate only re). In the latter case, Ke is calculated by ForwardDiff.jl

Dispatch is typically done on a user-defined material struct, and possible as well on cellvalues (potentially a NamedTuple/Tuple). State variables and current dof-values for the cell are directly available in the element_routine! Old dof-values for the cell, user-defined cache and cell_load types, cell coordinates and more are available through the CellBuffer type given as additional input.

A minimal example

The full example can be downloaded here.

First we create the dofhandler and cellvalues as in Ferrite.jl's heat equation example

using Ferrite, FerriteAssembly
dh = DofHandler(generate_grid(Quadrilateral, (20, 20))); push!(dh, :u, 1); close!(dh)
cellvalues = CellScalarValues(QuadratureRule{2, RefCube}(2), Lagrange{2, RefCube, 1}())

We start by defining the material (that normally contains material parameters but are hard-coded in the example)

struct ThermalMaterial end

and then define our element_routine! for that material as

function FerriteAssembly.element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state,
    ae::AbstractVector, material::ThermalMaterial, cellvalues::CellScalarValues,
    dh_fh, Δt, buffer::CellBuffer
    )
    n_basefuncs = getnbasefunctions(cellvalues)
    # Loop over quadrature points
    for q_point in 1:getnquadpoints(cellvalues)
        dΩ = getdetJdV(cellvalues, q_point)
        for i in 1:n_basefuncs
            δu  = shape_value(cellvalues, q_point, i)
            ∇δu = shape_gradient(cellvalues, q_point, i)
            # Add body load contribution to re
            re[i] += -δu * dΩ
            # Loop over trial shape functions
            for j in 1:n_basefuncs
                ∇u = shape_gradient(cellvalues, q_point, j)
                # Add contribution to Ke
                Ke[i, j] += (∇δu ⋅ ∇u) * dΩ
            end
        end
    end
end

which is basically the same as in Ferrite.jl's example.

We can now create the global vectors and matrices.

K = create_sparsity_pattern(dh)
r = zeros(ndofs(dh))

In general, each integration point (or cell if desired by the user) can have a state. In this case, we have no state variables, but the interface still requires them. We can then create states of nothing by

states = create_states(dh)

See create_states for more detailed options of creating actual state variables.

The next step is gathering all variables that are shared for all elements, but may be modified for or by each element in the cellbuffer:

cellbuffer = CellBuffer(dh, cellvalues, ThermalMaterial())

Note that cellvalues can be a Tuple or NamedTuple, this is useful for coupled problems with multiple fields.

We then define our assembler and do the assembly

assembler = start_assemble(K,r)
doassemble!(assembler, cellbuffer, states, dh)

Threaded assembly

To do the assembly in the example above threaded, we need to color the grid to avoid race conditions. This can be done with Ferrite.jl's create_coloring function:

colors = create_coloring(dh.grid)

We must also create threaded versions of the cellbuffer and assembler,

cellbuffers = create_threaded_CellBuffers(CellBuffer(dh, cellvalues, ThermalMaterial()))
assemblers = create_threaded_assemblers(K, r)

And then we can call doassemble! as

doassemble!(assemblers, cellbuffers, states, dh, colors)

The examples were generated using Literate.jl.

Detailed API description

Overloaded element routine

One of the element methods should be overloaded for a given combination of cellvalues and material. Note that the cellvalues are already reinit!:ed when passed to the element routines.

FerriteAssembly.element_routine! — Function

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state,
    ae::AbstractVector, material, cellvalues, 
    dh_fh::Union{DofHandler,FieldHandler}, Δt, buffer
    )

The main function to be overloaded for the specific material and cellvalues. This function should modify the element stiffness matrix Ke and the residual re.

state should contain a state description for the element. Typically, state will be a vector with a state variable for each integration point, but it can also be any other type for each element. On input, these are the old values and should be mutated to the updated value for the current time step and guess for ae.

The user defined material variable usually contain the material parameters.

cellvalues should contain the CellValues for the given element. It can also be a tuple or named tuple of cellvalues.

When the regular DofHandler is used, dh_fh::DofHandler is passed to the element routine. However, if the MixedDofHandler is used, one of its fieldhandlers are passed as dh_fh::FieldHandler. This gives the option to call dof_range(dh_fh, field::Symbol) for multi-field problems.

Δt is time increment given to doassemble

buffer is normally CellBuffer (if given to doassemble). Then, it can be used to get

buffer.ae_old: The old values of the displacements (if aold::Nothing is passed to doassemble, buffer.ae_old will be NaNs)

buffer.cell_load: The cell_load passed to CellBuffer, typically used for body loads or source terms.

buffer.cache: The cache passed to CellBuffer, typically used to gather all preallocations if such are necessary

getcoordinates(buffer)::Vector{Vec}: The cell's coordinates

celldofs(buffer)::Vector{Int}: The cell's global degrees of freedom numbers

FerriteAssembly.element_residual! — Function

element_residual!(
    re::AbstractVector, state, 
    ae::AbstractVector, material, cellvalues, 
    dh_fh::Union{DofHandler,FieldHandler}, Δt, buffer
    )

To calculate the element tangent stiffness Ke automatically by using ForwardDiff, it is possible to overload element_residual! instead of element_routine!. See element_routine! for a description of the input parameters.

Note that in order for this function to work, care must be taken when mutating values to not change their types. When mutating the state, ensure to call ForwardDiff.value() on those values. Warning: Only do this at a point when the calculation of re is unaffected by the values in state, otherwise Ke will be wrong.

CellBuffer

Variables that are used and modified for each cell of a certain type, but that don't belong to a specific cell, are collected in a CellBuffer.

FerriteAssembly.CellBuffer — Type

CellBuffer(
    numdofs::Int, numnodes::Int, ::Val{dim}, 
    cellvalues, material, cell_load=nothing, cache=nothing) -> CellBuffer

Create a cell cache for an element with numdofs degrees of freedom and numnodes nodes with dimension dim. Add the given cellvalues, material, and cache to the CellBuffer as well.

CellBuffer(
    dh::DofHandler, 
    cellvalues, material, cell_load=nothing, cache=nothing) -> CellBuffer

Use dh to get numdofs, numnodes, and dim, before calling the above method definition.

CellBuffer(
    dh::MixedDofHandler, fh::FieldHandler, 
    cellvalues, material, cell_load=nothing, cache=nothing) -> CellBuffer

Use dh and fh to get numdofs, numnodes, and dim, before calling the first CellBuffer method definition.

CellBuffer(
    dh::MixedDofHandler, cellvalues, 
    material, cell_load=nothing, cache=nothing) -> NTuple{N,<:CellBuffer}

Return a tuple of CellBuffers for each FieldHandler in dh.fieldhandlers. cellvalues[i] corresponds to dh.fieldhandlers[i], and so does materials[i], cell_load[i] and caches[i]. If only one CellValues, material, cell_load, and/or cache is given (not as a ::Tuple), the same is used for all fieldhandlers. If a tuple of cellvalues (or materials/cell_load/caches) should be used for each cell, and the same tuple should be used for each fieldhandler, then it must be given as a tuple of tuples. (Often, it is better to give a NamedTuple of e.g. CellValues to be used for every fieldhandler)

The Ferrite.jl functions getcoordinates(::CellBuffer) and celldofs(::CellBuffer) are defined and can be used inside an element routine to get the current cell's coordinates and dof-numbers.

State variables

The initial state variables may vary depending on the position in the grid. Furthermore, the datastructure depends on the type of dof handler, so a convenience function exists that creates the correct variable

FerriteAssembly.create_states — Function

create_states(dh::DofHandler, 
    statefun::Function=Returns(nothing), 
    cellvalues::Union{CellValues,Nothing}=nothing
    ) -> Vector

Creates a Vector of states for each cell in dh, where a state=statefun(x::Vec) and x is the coordinate in the grid. If isnothing(cellvalues), then the statefun is called once for the average nodal coordinate of each cell. Otherwise, it is called once per quadrature point in cellvalues for that cell.

create_states(dh::DofHandler, 
    statefun::Dict{String,Function}, 
    cellvalues::Union{Dict{String},CellValues,Nothing}}=nothing
    ) -> Dict{String,Dict{Int}}

create_states(dh::MixedDofHandler, 
    statefun::Dict{String,Function}, 
    cellvalues::Union{Dict{String},CellValues,Nothing}}=nothing
    ) -> Dict{String,NTuple{N,Dict{Int}}}

If a isa(statefun,Dict), then given a Dict{String,Dict{Int}} is returned, one for each cellset in the grid corresponding to the keys of statefun. The elements are then Dicts that map the cellnr to the state for each cell in the cellset. If cellvalues::Dict is given, the keys should match statefun. Otherwise, the same cellvalues are assumed for each statefun. This is typically used if multiple materials are used on the same grid.

create_states(dh::MixedDofHandler, 
    statefun::Union{Function,Tuple}=Returns(nothing), 
    cellvalues::Union{CellValues,Tuple,Nothing}=nothing,
    cellset=nothing
    ) -> NTuple{N,Dict{Int}}

Creates a Tuple of the output of the following function for each fieldhandler in dh. If statefuns and/or cellvalues are not tuples, the same value is used for each fieldhandler.

create_states(dh::MixedDofHandler, fh::FieldHandler, 
    statefun::Function, cellvalues::Union{Nothing,CellValues}, cellset
    ) -> Dict{Int}

Returns a Dict{Int} with states for each cell in fh's cellset and keys corresponding to the global cellid If isnothing(cellvalues), then the statefun(x::Vec) is called once for the average nodal coordinate x of each cell. Otherwise, it is called once for each quadrature point location x given by cellvalues for that cell. If a cellset!=nothing is given, states are only created for the intersection of that cellset and the fieldhandlers cellset

doassemble!

Once everything is set up, one can call the function which will actually do the assembly:

FerriteAssembly.doassemble! — Function

doassemble!(
    assembler::Ferrite.AbstractSparseAssembler, cellbuffer::CellBuffer, 
    s::AbstractVector, dh::DofHandler, 
    a=nothing, aold=nothing, Δt=nothing
    )

Sequential assembly of cells with the dh::DofHandler.

assembler is obtained from Ferrite.jl's start_assemble(K,r) function

cellbuffer contains buffers for the specific cell. See CellBuffer for more info.

s is a collection (vector, dict, etc.) of state variables, where indexing by cellnr gives the state variables for that cell.

a and aold are the current and old unknowns (can be set to nothing if not used)

Δt is the time increment passed into each element routine

doassemble!(
    assemblers::Vector{<:Ferrite.AbstractSparseAssembler},
    cellbuffers::Vector{<:CellBuffer}, states, 
    dh::DofHandler, colored_sets::Vector{Vector{Int}}, 
    a=nothing, aold=nothing, Δt=nothing
    )

Threaded assembly of cells with the dh::DofHandler.

assemblers are assemblers for each thread, which can be obtained with the create_threaded_assemblers function.

cellbuffers contains buffers for the specific cell, for each thread. This can be created by create_threaded_CellBuffers. See also CellBuffer for more info.

states, a, aold, and Δt are the same as for the sequential doassemble!

colored_sets are cellsets for each color

doassemble!(
    assembler::Ferrite.AbstractSparseAssembler, 
    cellbuffers::Tuple, states::Tuple, 
    dh::MixedDofHandler, 
    a=nothing, aold=nothing, Δt=nothing
    )

Sequential assembly of cells with the dh::MixedDofHandler.

cellbuffers contains buffers for the specific cell in each FieldHandler in dh.fieldhandlers See CellBuffer for more info.

states is a tuple which contains a collection of state variables, one vector for each FieldHandler in dh.fieldhandlers. Each element in the collection, i.e. states[cellnr], contains the state variables for one Cell with global number cellnr. Unless all cells have the same type of the state, it might make sense to use a Dict{Int,State} where the key refers to the global number.

assembler, a, aold, and Δt are the same as for the DofHandler case.

doassemble!(
    assemblers::Vector{<:Ferrite.AbstractSparseAssembler},
    cellbuffers::Tuple, 
    states::Tuple, 
    dh::MixedDofHandler, colored_sets::Vector{Vector{Int}}, 
    a::AbstractVector, aold::AbstractVector, Δt::Number
    )

Threaded assembly of cells with the dh::MixedDofHandler.

assemblers and colored_sets are the same as for the threaded DofHandler case.

states are the same as for the sequential MixedDofHandler case.

cellbuffers contains vectors Vector{CellBuffer} for the cell type in each FieldHandler in dh.fieldhandlers. The vector element corresponds to each thread. This can be created by create_threaded_CellBuffers. See also CellBuffer for more info.

a, aold, and Δt are the same as for the DofHandler case.

Threaded assembly

For parallel assembly, we need a vector of CellBuffers and assemblers, one for each thread.

For the MixedDofHandler, the outer loop is over the type of cells, so we need a tuple that contains vectors of CellBuffers. Construction of this via deepcopy is implemented as

FerriteAssembly.create_threaded_CellBuffers — Function

create_threaded_CellBuffers(c::CellBuffer; nthreads=Threads.nthreads())
create_threaded_CellBuffers(cs::Tuple; nthreads=Threads.nthreads())

Convenience function for creating a vector with cell buffers for each thread. The standard workflow is to first call CellBuffer with the dof handler. For DofHandler this will give a CellBuffer, and for MixedDofHandler this gives a tuple of CellBuffers. In both cases, the output can be given to create_threaded_CellBuffers to produce the appropriate result required by the threaded versions of doassemble!.

A vector of assemblers that is convieniently created by calling

FerriteAssembly.create_threaded_assemblers — Function

create_threaded_assemblers(K, r; nthreads=Threads.nthreads())

Convenience function for creating a nthreads long vector with the output of start_assemble as elements