AbstractItemBuffer

Depening on the domain that is worked over, different item buffers are available, e.g. CellBuffer and FaceBuffer. These are set up during call to setup_domainbuffer.

For each item (cell or face), the values in the buffer are updated to the current item, and can be accessed with the following functions:

FerriteAssembly.get_aeold — Function

get_aeold(itembuffer::AbstractItemBuffer)

Get the degrees of freedom pertinent to the values object (e.g. cellvalues) in the itembuffer.

Note: Filled by NaNs unless aold is passed to work!

FerriteAssembly.get_old_state — Function

get_old_state(itembuffer::AbstractCellBuffer)

Get the old state variables for the cell. Currently only available for cells and not for faces.

FerriteAssembly.get_time_increment — Function

get_time_increment(itembuffer::AbstractItemBuffer)

Get the time increment to get to the current step. Set by set_time_increment!.

Ferrite.dof_range — Method

Ferrite.dof_range(::AbstractItemBuffer, ::Symbol)

Get the dof_range for a specific field, same as Ferrite.dof_range(::SubDofHandler, ::Symbol), but fully type-stable.

Ferrite.getcoordinates — Method

Ferrite.getcoordinates(::AbstractItemBuffer)

Get the cell coordinates for the current item.

Ferrite.celldofs — Method

Ferrite.celldofs(::AbstractItemBuffer)

Get the degree of freedom indices for the current cell referred to by the current item.

Ferrite.cellid — Method

Ferrite.cellid(::AbstractItemBuffer)

Get the cell nr for the current item.

FerriteAssembly.get_user_data — Function

get_user_data(itembuffer::AbstractItemBuffer)

Get the user_data passed to the DomainSpec when setting up the domain. This is not modified and always passed around as reference.

FerriteAssembly.get_user_cache — Function

get_user_cache(itembuffer::AbstractItemBuffer)

Get the user_cache created by allocate_cell_cache. For multithreaded applications, this cache is copied between for each tasks, and can be modified without risking race conditions.

AbstractCellBuffer

Two types of cell buffers are provided, the regular CellBuffer and a wrapper AutoDiffCellBuffer that speeds up the automatic differentiation. By using the above access functions, these behave identical.

To allocate the user cache, overload allocate_cell_cache:

FerriteAssembly.allocate_cell_cache — Method

FerriteAssembly.allocate_cell_cache(material, cellvalues)

This function can be overloaded for the specific material to allocate a cache that is stored in the AbstractCellBuffer. This cache can be used to reduce allocations. Returns nothing by default.

FaceBuffer

The FaceBuffer is similar to the CellBuffer, except that it has FaceValues instead of CellValues and that state variables are not supported. To allocate user cache, overload allocate_face_cache:

FerriteAssembly.allocate_face_cache — Function

FerriteAssembly.allocate_face_cache(material, cellvalues)

This function can be overloaded for the specific material to allocate a cache that is stored in the FaceBuffer and can be used to reduce allocations. Returns nothing by default.