State variables

The state variable for a given cell is determined by the material type, via overloading the create_cell_state function. To update old states to the new states, use update_states!.

The state variable datastructure

The state variables are created when calling setup_domainbuffer or setup_domainbuffers, and stored inside the buffers. The states for a given domain are accessed with get_state and get_old_state, where the state for a particular cell is indexed by its cell number. (The output from the mentioned functions are Dict{Int})

API

FerriteAssembly.create_cell_state — Method

create_cell_state(material, cellvalues, x, ae, dofrange)

Defaults to returning nothing.

Overload this function to create the state which should be passed into the element_routine!/element_residual! for the given material and cellvalues. x is the cell's coordinates, ae the element degree of freedom values, and dofrange::NamedTuple containing the local dof range for each field. As for the element routines, ae, is filled with NaN unless the global degree of freedom vector is given to the setup_domainbuffer function.

FerriteAssembly.update_states! — Method

update_states!(old_states, states)

In most cases, this 2-argument function is not required, and the entire domain buffer can be passed instead, see update_states!. This 2-argument function will then be called for the stored state variables.