Example elements

The package includes a set of example elements inside the submodule FerriteAssembly.ExampleElements:

FerriteAssembly.ExampleElements.FourierMaterial — Type

FourierMaterial(k)

For solving the linear heat equation (which uses Fourier's law) according to the strong form,

\[ -\nabla \cdot (k \nabla T) = f \quad \textbf{x} \in \Omega,\]

where $T$ is the temperature field, $k$ the thermal conductivity, and $f$ a volume source term. $f$ is not included in the element assembly, and can be applied using FerriteNeumann.jl Only the stiffness contribution is assembled.

FerriteAssembly.ExampleElements.TransientFourierMaterial — Type

TransientFourierMaterial(k, c)

For solving transient heat flow, assuming flux according to Fourier's law with conductivity, k, and heat capacity, c, according to

\[ c\frac{\partial T}{\partial t}-\nabla \cdot (k \nabla T) = f \quad x \in \Omega,\]

where $T$ is the temperature field and $f$ is a volume source term. $f$ is not included in the element assembly, but can be included with FerriteNeumann.jl. Both the stiffness matrix, $K=\mathrm{d}f_{int}/\partial T$ and the residual vector, $r=f_{int}-f_{ext}$ are assembled.