Setup

To setup the assembly, call setup_assembly

For multiple domains, AssemblyDomains must be created first.

FerriteAssembly.setup_assembly — Function

setup_assembly(dh, material, cellvalues; kwargs...)

Setup assembly for a single domain (i.e. the same material and interpolations everywhere). Returns the buffer, old_states, and new_states to be used in doassemble!. The state variables are created by calling the user-defined create_cell_state function. Available keyword arguments

a=nothing: Give the global dof vector to pass element dofs, ae, to create_cell_state (NaN-values otherwise)

threaded=Val(false): Set to Val(true) to setup threaded assembly. More elaborate settings may be added in the future.

autodiffbuffer=Val(false): Set to true or Val(true) (for type stable construction) to use AutoDiffCellBuffer instead of CellBuffer.

user_data=nothing: The user_data is passed to each AbstractCellBuffer by reference (when threaded)

cache=nothing: The cache is passed to each AbstractCellBuffer, and deepcopied if threaded.

colors=nothing: Give colors for the grid from Ferrite.create_coloring to setup threaded colored assembly. If nothing, Ferrite's default coloring algorithm is used.

cellset="all cells in dh": Which cells to assemble. In most cases, it is better to setup AssemblyDomains to assemble different domains. But this option can be used to assemble only a subset of the grid.

setup_assembly(domains::Vector{<:AssemblyDomain}; kwargs...)

Setup assembly for each AssemblyDomain in domains. Returns the Dicts buffers, old_states, and new_states to be used in doassemble!. The state variables are created by calling the user-defined create_cell_state function. Available keyword arguments

a=nothing: Give the global dof vector to pass element dofs, ae, to create_cell_state (NaN-values otherwise)

autodiffbuffer=Val(false): Set to true or Val(true) (for type stable construction) to use AutoDiffCellBuffer instead of CellBuffer.

threading=Val(false): Set to Val(true) to setup threaded assembly. More elaborate settings may be added in the future.

FerriteAssembly.AssemblyDomain — Type

AssemblyDomain(name, dh, material, cellvalues; cellset, colors=nothing, user_data=nothing, cache=nothing)

Create an AssemblyDomain that can be used when calling setup_assembly to assemble multiple domains. name is used to access the corresponding DomainBuffer and state variables returned by setup_assembly. If not given, cellset is attempted to be inferred from the DofHandler, dh.