Workers

There are currently two categories of workers implemented:

Assemblers: Calculate system matrices and vectors

Integrators: Integrate a function over the domain

Perform the work

Having setup a DomainBuffer and a worker, work is performed by calling the function work!

FerriteAssembly.work! — Function

work!(worker, buffer; a=nothing, aold=nothing)

Perform the work according to worker over the domain(s) specificed by buffer. Current, a, and old, aold, global degree of freedom vectors are passed to get these values passed into the innermost user-defined functions. If not passed (or as nothing), NaN values are passed into the innermost functions.

How a worker works

work_single_cell! work_single_face!

Support threading

TaskLocals interface + can_thread defined