Workers

There are currently three categories of workers implemented:

Assemblers: Calculate system matrices and vectors

Integrators: Integrate a function over the domain

QuadPointEvaluator: Evaluate a function in each quadrature point

Perform the work

Having setup a DomainBuffer and a worker, work is performed by calling the function work!

FerriteAssembly.work! — Function

work!(worker, buffer; a=nothing, aold=nothing)

Perform the work according to worker over the domain(s) specified by buffer. Current, a, and old, aold, global degree of freedom vectors are passed to get these values passed into the innermost user-defined functions. If not passed (or as nothing), NaN values are passed into the innermost functions.

A worker needs to define what to do for different domains, specifically it needs to define work_single_cell! and work_single_facet!. Further details about writing custom workers can be found in the Customizations section.