API

Methods that should be overloaded and exported functions are described on this page. It may also be useful to checkout the Data structures.

Element routines

One of the element methods should be overloaded for a given combination of cellvalues and material. Note that the cellvalues are already reinit!:ed when passed to the element routines.

FerriteAssembly.element_routine! — Function

element_routine!(
    Ke::AbstractMatrix, re::AbstractVector, state,
    ae::AbstractVector, material, cellvalues, 
    dh_fh, Δt, buffer
    )

The main function to be overloaded for the specific material and cellvalues. This function should modify the element stiffness matrix Ke and the residual re.

state should contain a state description for the element. Typically, state will be a vector with a state variable for each integration point, but it can also be any other type for each element. On input, these are the old values and should be mutated to the updated value for the current time step and guess for ae.

The user defined material variable usually contain the material parameters.

cellvalues should contain the CellValues for the given element. It can also be a tuple or named tuple of cellvalues.

When the regular DofHandler is used, dh_fh::DofHandler is passed to the element routine. However, if the MixedDofHandler is used, one of its fieldhandlers are passed as dh_fh::FieldHandler. This gives the option to call dof_range(dh_fh, field::Symbol) for multi-field problems. Note: Please do not rely on dh_fh for anything but dof_range, as dh_fh may be replaced with another type that only supports dof_range.

Δt is time increment given to doassemble

buffer::Union{CellBuffer, AutoDiffCellBuffer} can be used to get

getcoordinates(buffer)::Vector{Vec}: The cell's coordinates

celldofs(buffer)::Vector{Int}: The cell's global degrees of freedom numbers

FerriteAssembly.get_aeold(buffer): aold[celldofs] (if aold::Nothing is passed to doassemble, a vector, [NaN for d in celldofs] is returned)

FerriteAssembly.get_load(buffer): The cell_load in the buffer, typically used for body loads or source terms.

FerriteAssembly.get_cache(buffer): The cache in the buffer - typically used to gather all preallocations if such are necessary

FerriteAssembly.element_residual! — Function

element_residual!(
    re::AbstractVector, state, 
    ae::AbstractVector, material, cellvalues, 
    dh_fh, Δt, buffer
    )

To calculate the element tangent stiffness Ke automatically by using ForwardDiff, it is possible to overload element_residual! instead of element_routine!. See element_routine! for a description of the input parameters.

Note that in order for this function to work, care must be taken when mutating values to not change their types. When mutating the state, ensure to call ForwardDiff.value() on those values. Warning: Only do this at a point when the calculation of re is unaffected by the values in state, otherwise Ke will be wrong.

CellBuffer

Variables that are used and modified for each cell of a certain type, but that don't belong to a specific cell, are collected in a CellBuffer.

FerriteAssembly.CellBuffer — Type

CellBuffer(
    numdofs::Int, numnodes::Int, ::Val{dim}, 
    cellvalues, material, cell_load=nothing, cache=nothing) -> CellBuffer

Create a cell cache for an element with numdofs degrees of freedom and numnodes nodes with dimension dim. Add the given cellvalues, material, and cache to the CellBuffer as well. Note that this constructor is normally not used, and is instead called from setup_cellbuffer.

FerriteAssembly.setup_cellbuffer — Function

setup_cellbuffer(
    dh::DofHandler, cellvalues, material, 
    cell_load=nothing, cache=nothing)

Creates a single CellBuffer for use with the standard DofHandler and a single material.

setup_cellbuffer(
    dh::MixedDofHandler, cellvalues, material, 
    cell_load=nothing, cache=nothing)

Return a tuple of CellBuffers for each FieldHandler in dh.fieldhandlers. cellvalues[i] corresponds to dh.fieldhandlers[i], and so does materials[i], cell_load[i] and caches[i]. If only one CellValues, material, cell_load, and/or cache is given (not as a ::Tuple), the same is used for all fieldhandlers. If a tuple of cellvalues (or materials/cell_load/caches) should be used for each cell, and the same tuple should be used for each fieldhandler, then it must be given as a tuple of tuples. (Often, it is better to give a NamedTuple of e.g. CellValues to be used for every fieldhandler)

The Ferrite.jl functions getcoordinates(::CellBuffer) and celldofs(::CellBuffer) are defined and can be used inside an element routine to get the current cell's coordinates and dof-numbers.

setup_cellbuffer(dh::AbstractDofHandler, cv, materials::Dict, 
    cell_load=nothing, cache=nothing)

Return a Dict{String} for each material in materials. If any of cv, cell_load, or cache is not a Dict, the same value is used for each CellBuffer. If Dicts are used, the keys must match those in materials. The return type depends on the dh:

dh::DofHandler: Dict{String,<:CellBuffer}

dh::MixedDofHandler: Dict{String,NTuple{N,CellBuffer}

FerriteAssembly.AutoDiffCellBuffer — Type

AutoDiffCellBuffer(cb::CellBuffer, cellstates, dh_fh::Union{DofHandler,FieldHandler})

cellstates should be the states for the relevant cells in dh_fh. An AutoDiffCellBuffer wraps a CellBuffer and behaves the same way wrt defined functions, i.e. all get* functions defined for CellBuffer are also defined for AutoDiffCellBuffer. When used for automatic differentiation, only the wrapped CellBuffer is passed to element_residual!. As a backup solution if direct field access is used, the function getCellBuffer returns the CellBuffer for both a CellBuffer and an AutoDiffCellBuffer.

Note that this constructor is normally not used, but called from the setup_ad_cellbuffer function.

FerriteAssembly.setup_ad_cellbuffer — Function

setup_ad_cellbuffer(states, dh::AbstractDofHandler, args...)

Create a special cell buffer that improves performance when using automatic differentiation to calculate the element stiffness. The exact same inputs as for setup_cellbuffer should be used, except that the states variable that is passed to doassemble! should be prepended to the argument list.

FerriteAssembly.getCellBuffer — Function

getCellBuffer(b::Union{CellBuffer,AutoDiffCellBuffer})

Return the CellBuffer if isa(b,CellBuffer), else, return the CellBuffer wrapped by b::AutoDiffCellBuffer, i.e. b.cb

Access functions

The following access functions can be used to extract information from the CellBuffer

Ferrite.getcoordinates — Function

Ferrite.getcoordinates(c::CellBuffer)

Ferrite.jl's getcoordinates function is overloaded on the CellBuffer to return the current cell's nodal coordinates.

FerriteAssembly.get_aeold — Function

FerriteAssembly.get_aeold(c::CellBuffer)

Get the old element dof-values for the current cell

FerriteAssembly.get_load — Function

FerriteAssembly.get_load(c::CellBuffer)

Get the user specified body load given to CellBuffer

FerriteAssembly.get_cache — Function

FerriteAssembly.get_cache(c::CellBuffer)

Get the user-specified cache given to the CellBuffer

State variables

The initial state variables may vary depending on the position in the grid. Furthermore, the datastructure depends on the type of dof handler, so a convenience function exists that creates the correct variable structure.

FerriteAssembly.create_states — Function

create_states(dh::DofHandler, material=nothing, cellvalues=nothing, a=nothing)

Create the state variables for the given dh, material, cellvalues, and global dof vector a. material and cellvalues can be a Dict{String}, in which case they should have the same keys, corresponding to cellsets in the grid. This follows the same structure as for creating a CellBuffer. The function create_cell_state is called for each cell.

create_states(dh::MixedDofHandler, material=nothing, cellvalues=nothing, a=nothing)

Create the state variables for the given dh, material, cellvalues, and global dof vector a.

material and cellvalues follow the same input structure as to setup_cellbuffer. I.e., they can be tuples with the same length as fh.fieldhandlers. Note that if cellvalues::Tuple is desired to be passed into the create_cell_state function, it must be wrapped in an outer ntuple with the same length as fh.fieldhandlers. (But using a NamedTuple is recommended to avoid this problem).

For multiple materials, material::Dict{String}, and cellvalues can optionally be that as well, but then they must have the same keys. This follows the same structure as for creating a CellBuffer. The function create_cell_state is called for each cell.

The order of each call to create_cell_state is fixed for a given input, i.e. the loops are only over sorted Sets and Dict keys. (Useful if using StableRNGs.jl for random variations to obtain reproducible results)

FerriteAssembly.create_cell_state — Function

create_cell_state(material, cellvalues, x, ae)

Defaults to returning nothing.

Overload this function to create the state which should be passed into the element_routine!/element_residual! for the given material and cellvalues. As for the element routines, ae, is filled with NaN unless the global degree of freedom vector is given to the create_states function.

doassemble!

Once everything is set up, one can call the function which will actually do the assembly:

FerriteAssembly.doassemble! — Function

doassemble!(
    assembler::Ferrite.AbstractSparseAssembler, cellbuffer::AbstractCellBuffer, 
    s::AbstractVector, dh::DofHandler, 
    a=nothing, aold=nothing, Δt=nothing, scaling=nothing; cellset=1:ncells
    )

Sequential assembly of cells with the dh::DofHandler.

assembler is obtained from Ferrite.jl's start_assemble(K,r) function

cellbuffer contains buffers for the specific cell. See CellBuffer for more info.

s is a collection (vector, dict, etc.) of state variables, where indexing by cellnr gives the state variables for that cell.

a and aold are the current and old unknowns (can be set to nothing if not used)

Δt is the time increment passed into each element routine

scaling allows including element data in order to scale the residuals, see Residual scaling

doassemble!(
    assemblers::Vector{<:Ferrite.AbstractSparseAssembler},
    cellbuffers::Vector{<:AbstractCellBuffer}, states, 
    dh::DofHandler, colored_sets::Vector{Vector{Int}}, 
    a=nothing, aold=nothing, Δt=nothing, 
    scalings::Vector=create_threaded_scalings(nothing);
    cellset=nothing
    )

Threaded assembly of cells with the dh::DofHandler.

assemblers are assemblers for each thread, which can be obtained with the create_threaded_assemblers function.

cellbuffers contains buffers for the specific cell, for each thread. This can be created by create_threaded_CellBuffers. See also CellBuffer for more info.

states, a, aold, and Δt are the same as for the sequential doassemble!

colored_sets are cellsets for each color

doassemble!(
    assembler::Ferrite.AbstractSparseAssembler, 
    cellbuffers::Tuple, states::Tuple, 
    dh::MixedDofHandler, 
    a=nothing, aold=nothing, Δt=nothing, scaling=nothing;
    kwargs...
    )

Sequential assembly of cells with the dh::MixedDofHandler.

cellbuffers contains buffers for the specific cell in each FieldHandler in dh.fieldhandlers See CellBuffer for more info.

states is a tuple which contains a collection of state variables, one vector for each FieldHandler in dh.fieldhandlers. Each element in the collection, i.e. states[cellnr], contains the state variables for one Cell with global number cellnr. Unless all cells have the same type of the state, it might make sense to use a Dict{Int,State} where the key refers to the global number.

assembler, a, aold, and Δt are the same as for the DofHandler case.

doassemble!(
    assemblers::Vector{<:Ferrite.AbstractSparseAssembler},
    cellbuffers::Tuple, 
    states::Tuple, 
    dh::MixedDofHandler, colored_sets::Vector{Vector{Int}}, 
    a::AbstractVector, aold::AbstractVector, Δt::Number, 
    scalings::Vector=create_threaded_scalings(nothing);
    kwargs...)

Threaded assembly of cells with the dh::MixedDofHandler.

assemblers and colored_sets are the same as for the threaded DofHandler case.

states are the same as for the sequential MixedDofHandler case.

cellbuffers contains vectors Vector{AbstractCellBuffer} for the cell type in each FieldHandler in dh.fieldhandlers. The vector element corresponds to each thread. This can be created by create_threaded_CellBuffers. See also CellBuffer for more info.

a, aold, and Δt are the same as for the DofHandler case.

Threaded assembly

For parallel assembly, we need a vector of CellBuffers and assemblers, one for each thread.

For the MixedDofHandler, the outer loop is over the type of cells, so we need a tuple that contains vectors of CellBuffers. Construction of this via deepcopy is implemented as

FerriteAssembly.create_threaded_CellBuffers — Function

create_threaded_CellBuffers(c::CellBuffer; nthreads=Threads.nthreads())
create_threaded_CellBuffers(cs::Tuple; nthreads=Threads.nthreads())
create_threaded_CellBuffers(cs::Dict{String}; nthreads=Threads.nthreads())

Convenience function for creating a vector with cell buffers for each thread. The standard workflow is to first call CellBuffer with the dof handler. For DofHandler this will give a CellBuffer, and for MixedDofHandler this gives a tuple of CellBuffers. In both cases, the output can be given to create_threaded_CellBuffers to produce the appropriate result required by the threaded versions of doassemble!. Similarily for a grid with mixed materials created by a dictionary of materials, just pass the created CellBuffer to this function, and the output is what doassemble! expects.

A vector of assemblers that is convieniently created by calling

FerriteAssembly.create_threaded_assemblers — Function

create_threaded_assemblers(K, r; nthreads=Threads.nthreads())

Convenience function for creating a nthreads long vector with the output of start_assemble as elements

Residual scaling

There are many options for how to scale the residual in finite element simulations. This package does not intend to implement many different options, but does give the user the option to calculate scaling contributions from each cell, which may be useful. By defining a scaling that is passed to doassemble!, this can be updated in each cell. One type of scaling, ElementResidualScaling, is included as described below. Its code can be used as a template for how to include custom scaling that works on the element level.

In general, a scaling must support the update_scaling! function. For consistency it is also nice, but not required, to support reset_scaling!. This function must, however, be called by the user before assembly (to allow consistent separated assembly using different cellsets).

FerriteAssembly.update_scaling! — Function

update_scaling!(scaling, re, dh_fh, cellbuffer)

This function should add the contribution from the element residual vector re to the scaling factors in scaling. The cellbuffer is included, which contains extra information if needed.

FerriteAssembly.reset_scaling! — Function

reset_scaling!(scaling)

This function should reset the scaling factors, such that the values don't accumulate if used in multiple iterations/time steps.

Additionally, create_threaded_scalings can be used to copy one scaling to each thread when using with parallel assembly.

FerriteAssembly.create_threaded_scalings — Function

create_threaded_scalings(scaling; nthreads=Threads.nthreads()) -> Vector

Makes nthreads deepcopies of scaling, such that it can be used for threaded assembly. Note: If creating your own scaling, it might make sense to define Base.sum(::Vector{<:MyScalingType}) to sum scalings after threaded assembly.

ElementResidualScaling

FerriteAssembly.ElementResidualScaling — Type

ElementResidualScaling(dh::AbstractDofHandler, p=2, T=Float64)

Create tolerance scaling based on the sum of the p-norm of each cell's residual vector, separately for each field. I.e. pseudo-code for field :u the scaling factor::T is

for each cell
    element_routine!(Ke, re, args...) # Calculate re
    factor += sum(abs.(re[dof_range(dh, :u)]).^p)^(1/p)