Internal API

Note that the internal API may change without being considered a breaking change!

FerriteAssembly.assemble_cell! — Function

assemble_cell!(assembler, cellbuffer, dh::DofHandler, cellnr, a, aold, state, Δt)
assemble_cell!(assembler, cellbuffer, dh::MixedDofHandler, fh::FieldHandler, cellnr, a, aold, state, Δt)

Internal function to that reinitializes the cellbuffer and calls assemble_cell_reinited!.

FerriteAssembly.inner_doassemble! — Function

inner_doassemble!(
    assembler, cellbuffer::AbstractCellBuffer, states, 
    dh::MixedDofHandler, fh::FieldHandler, a, aold, Δt
    )

Sequential assembly of cells corresponding to the given fh from dh.fieldhandlers. Internal function that is called from the sequential version of doassemble! for the MixedDofHandler

inner_doassemble!(
    assemblers::Vector{<:Ferrite.AbstractSparseAssembler},
    cellbuffers::Vector{<:AbstractCellBuffer}, states, 
    dh::MixedDofHandler, fh::FieldHandler, 
    colored_sets::Vector{Vector{Int}}, 
    a, aold, Δt)

Parallel assembly of cells corresponding to the given fh from dh.fieldhandlers. Internal function that is called from the parallel version of doassemble! for the MixedDofHandler

FerriteAssembly.assemble_cell_reinited! — Function

assemble_cell_reinited!(assembler, cellbuffer, dh_fh::Union{DofHandler,FieldHandler}, state, Δt)

Internal function that assembles the cell described by the reinitialized cellbuffer. This function is called in all cases: Parallel or sequential and DofHandler or MixedDofHandler

FerriteAssembly._copydofs! — Function

_copydofs!(edofs::Vector, gdofs::Vector, inds::Vector{Int})

Internal function for faster copying of global values into the element values. Equivalent to edofs .= gdofs[inds] _copydofs!(edofs::Vector, gdofs::Nothing, inds::Vector{Int})

Fill edofs with NaN

FerriteAssembly._maketuple — Function

_maketuple(t, n)

If t is a tuple, check that length(t)=n and return t. Otherwise, return a tuple of length n with t as every element

FerriteAssembly._makedict — Function

_makedict(d, d_keys)

If d::Dict then check that it has all keys in d_keys (if not throw KeyError). Otherwise, return a Dict with keys d_keys and d as every element

FerriteAssembly.create_state — Function

create_state(statefun, cell, cv::Nothing)
create_state(statefun, cell, cv::CellValues)

Internal functions for calling statefun(x) for the cell center if cv::Nothing or for each quadrature point if cv::CellValues.

Ferrite.reinit! — Function

Ferrite.reinit!(c::CellBuffer, dh::AbstractDofHandler, cellnum::Int, anew, aold)

Reinitialize the c::CellBuffer for cell number cellnum. The global degree of freedom vectors anew (current) and aold are used to update the cell degree of freedom vectors in c. If the global vectors are instead ::Nothing, the corresponding cell values are set to NaN The element stiffness, c.Ke, and residual, c.re, are also zeroed.

FerriteAssembly.intersect_nothing — Function

intersect_nothing(a, b)

A faster intersect if cellset::Nothing, otherwise calls Base.intersect