Internal API

Note that the internal API may change without being considered a breaking change!

FerriteAssembly.create_states — Function

create_states(sdh::SubDofHandler, material, cellvalues, a, cellset, dofrange)

Create a Dict of states for the cells in cellset, where the user should define the create_cell_state function for their material (and corresponding cellvalues) dofrange::NamedTuple is passed onto create_cell_state and contains the local dof ranges for each field.

FerriteAssembly.assemble_cell! — Function

assemble_cell!(assembler, new_states, cellbuffer, sdh::SubDofHandler, cellnr, a, aold, old_states)

Internal function to that reinitializes the cellbuffer and calls assemble_cell_reinited!.

FerriteAssembly._copydofs! — Function

_copydofs!(edofs::Vector, gdofs::Vector, inds::Vector{Int})

Internal function for faster copying of global values into the element values. Equivalent to edofs .= gdofs[inds]

_copydofs!(edofs::Vector, gdofs::Nothing, inds::Vector{Int})

Fill edofs with NaN

Ferrite.reinit! — Function

Ferrite.reinit!(c::CellBuffer, dh::AbstractDofHandler, cellnum::Int, anew, aold)

Reinitialize the c::CellBuffer for cell number cellnum. The global degree of freedom vectors anew (current) and aold are used to update the cell degree of freedom vectors in c. If the global vectors are instead ::Nothing, the corresponding cell values are set to NaN The element stiffness, c.Ke, and residual, c.re, are also zeroed.

FerriteAssembly._create_cell_state — Function

_create_cell_state(cell, material, cellvalues, a, ae, dofrange, cellnr)

Internal function to reinit and extract the relevant quantities from the cell::CellCache, reinit cellvalues, update ae from a, and pass these into the create_cell_state function that the user should specify.

FerriteAssembly.CellBuffer — Type

CellBuffer(
    numdofs::Int, numnodes::Int, ::Val{sdim}, 
    cellvalues, material, cell_load=nothing, cache=nothing) -> CellBuffer

Create a cell cache for an element with numdofs degrees of freedom and numnodes nodes with dimension sdim. cellvalues are reinit!ed for each cell, and the state is updated to the old cell state. material will be passed as-is to the element. The given dofrange::NamedTuple, user_data::Any, and cache::Any are available to the element via the buffer input.

See setup_assembly

This constructor is normally not used, and is instead called from setup_assembly

FerriteAssembly.AutoDiffCellBuffer — Type

AutoDiffCellBuffer(cb::CellBuffer)

FerriteAssembly.fast_getindex — Function

fast_getindex(collection)

If the output is known from the element type of the collection, this can be returned directly. Currently, this is implemented for AbstractDict with Nothing as type, which takes away overhead when state variables are not used.